Increase leveldb max memory to 70
Reducing to 30 didn't seem to have a meaningful impact on memory consumption because the cluster we were using it on only had 1GB of RAM per node, which is not enough for a ring size of 128. We have since increased the RAM to 4GB per node and the cluster has been stable since then.