make run-shell-command --become and --become-user work intuitively
run-shell-command will now automatically determine whether it needs to
include the vault file (and thus ask for the vault password) in order to
run a command, based on the vaules of --become and --become-user. If you
are using --become, you need the vault file because it includes the
ansible user's password; the exception is that if you are doing
something as cchq, you do *not* need a password, because the ansible
user is able to execute things as cchq without a password.

This opens up a workflow where you can run something as cchq without a
password, and if that fails for permissions reasons, you can fall back
to running as root and giving the ansible password.