make riakcs new use set_site_present
a very careful reading of this diff will reveal that there is a small
change in the logic for removal of this site. It used to be that in order
for it to be removed, [riakcs] group had to exist and not be empty.
Not if it does _not_ exist or _is_ empty, that is a condition under
which it _will_ be removed. This is unusual, but I think techincally correct:
if we move from previously using riakcs to not (e.g. because we move to Amazon S3)
then we will want to remove this as well I think.