Add terraform-migrate-state command
and beginnings of a state migration system for terraform.

Terraform assigns each resource an address so that it can map it back to the code.
However, often when you change the code, the addresses no longer map to the same place.
For this, terraform offers the `terraform state mv <address> <new_address>` command,
so you can tell it how existing resources map to your new code.

This is a tedious task, and often follows a very predictable renaming pattern.
So, this change introduces the beginnings of a system for specifying the renaming rule
and then having the state migrations automated, to reduce tedium and human error.