Give nginx_sites tag its intended purpose back
a `tags` on `include_roles` just tags the _include_role_ task,
not any of the included tasks. To tag all of the included tasks (recursively),
you have to _also_ tag it in the `name` arg. (_Only_ tagging it here does not
work either; you have to tag it in both places.)