Add postgresql params to pgsynclog0-production
Other RDS instances aren't one-to-one with current prod,
so I want to give RDS's "sensical" default values a chance before trying
to manually set them to something. I did want to prove that this works
on at least one db so that if we realize in a firefighting situation
that we need to set these values, we don't have to write the code to
support it on the fly.