Make COUCH_REINDEX_SCHEDULE less frequent on production
I decreased this yesterday and things were _worse_ not better last night.
Though it may be unrelated, I am going to try taking this in the opposite direction.
It's possible such frequent reindexing caused unnecesary churn
and CPU spent on rarely accessed views when it was needed to process real requests.