Make pgbouncer_hba reference pgbouncer_host
The current template was relying on the fact that on some environments (e.g. ICDS, india)
pgbouncer_host is always the same as host. On staging and production, which use Amazon RDS,
that becomes a problem. Not sure exactly when this took effect, but it became an issue
when this template was used to build a new pgproxy machine on staging, which would not allow
connections from any of machines except for the commcarehq_proxy db.