Fix the code for deleting unneeded supervisor confs
This code was badly broken. For one thing, it was iterating over items that were made up of the
single literal string "contents.stdout_lines",
instead of the value of the expression `contents.stdout_lines`.
Once that was fixed, I discovered a number of other issues
that needed to be fixed.