run postgresql_base as include instead of dependency
Making it a dependency means that the citus extension is configured
in the postgresql.conf file before it is installed.

Adding the extension later means that when the role is re-run
the extension is removed from postgresql.conf and then re-added later
with a restart in between which causes errors.

Doing it this way we can install the citus extension and then run
the postgresql_base role.