Fix serious formplayer issue after deploy with stop-gap measure
My temporary fix is to just restart formplayer after deploy,
even though it wasnt actually updated. And since there's a restart there
will still be the 1-2m of formplayer downtime that we're "used" to. The
longer term fix will be to move the formplayer directory out from inside
the commcare-hq code directory so that deploying commcare doesn't change
the file pointed to by the formplayer jar path that supervisor uses to run the jar.