create virutalenv using release path not 'current'
When creating a new virutalenv with an absolute path the path
will not be resolved if it is a link. Instead virtualenv will
just use the path as is.

This break virtualenv-clone since it looks for instances of 'old_dir'
to replace with 'new_dir'. 'old_dir' is the actual path (not link) which it
does not find since the path that is in the files is the 'link' path.