Undo recent changes to ask
I think I was thinking about this the wrong way, and I decided I don't
actually want to touch this. I realized we rely on ask having a
particular meaning: it means "here's something that you should say yes to
if you want to do the normal thing". To present a consistent experience,
if we want the default to be no, we should just ask the question in the
negative: "Do you want to skip formplayer deploy? [Y/n]". This allows
us to always pick a predictable answer when using --quiet, which means "say
yes to everything". In the previous commit where I changed quiet to
always pick the default, that was obviously a problem because our
default is current No everywhere, but we want quiet to pick Yes for you.