Fix virtualenvs that ref each other (via current)
This is essentially the same issue as https://github.com/dimagi/commcare-cloud/pull/3191
but solves it more fully. Scripts in virtualenv bin/ all reference the
current virtualenv in their #! line
(if the machine was set up while the issue in that PR was live)
except for activate which that PR's changelog entry fixes.
This resulted in _all_ virtualenvs' pip script pointing to the current
one, so pip install in any virtualenv actually pip installed in current,
making all virtualenvs essentially all act like one virtualenv. This is
obviously wrong and dangerous for multiple reasons, but we discovered it
because a side effect is that the deploy updates current instead of the
new virtualenv and then promotes the new virtualenv (which hasn't
received new pip installs) to current, resulting in libraries not
getting updated when they should.