Route by new formplayer_session cookie instead of sessionid cookie
Currently when a user logs out and logs back in, they are routed
randomly to one of the formplayer machines. This change gives the caller
the freedom to route based on whatever they like, and we will use that
in commcare-hq to route based on user_id instead of django sessionid.

The caller can now route by setting either the X-FORMPLAYER-SESSION
header or the formplayer_session cookie. A blemish of this
implementation is that if they sometimes but not always send _both_ the
header and the cookie, that will break sticky routing.