Simplify pull request template
I took a look at the 50 most recent PRs.  None included anything in the "ADDITIONAL INFORMATION" section.  Three included a "COMPONENT NAME", but that name was also in the PR title.  Where "ISSUE TYPE" was filled out, it was only ever "Change Pull Request" (not especially useful information).  "ENVIRONMENTS AFFECTED" was usually filled out, and that does seem useful to have.

Removing the rest of this since it's an annoyance everytime I make a new PR.