Allow --commcare-rev with --set on deploy
Previously if you use --set, it will effectively overwrite the value
of --commcare-rev (since that also uses --set on fab downstream)

This change also removes some of the awkwardness of using --set
where you have to then explictly pass in the component arg or it
treats the value of --set as the component arg (and fails).
Now there is none of that.