Enable commcare-cloud --control with python3
Set the environment variable CCHQ_VIRTUALENV to 'cchq' (or anything other than 'ansible') when running commcore-cloud --control commands to cause the virtualenv with that name to be used on the control machine. A virtualenv with will be created with python3 if it does not exist. Example:

CCHQ_VIRTUALENV=cchq commcare-cloud <env> --control ...

Use of virtualenvwrapper by init.sh will be phased out with Python 2.