Increase pgucr0-production RDS instance to make use of RIs
It hasn't been increased this year while the rest of the cluster expanded around it.
It's the only db.t3 we use in production and since the maintainance in the last couple weeks
we are no longer using all of our RIs. Based on those two things it seemed worth it to throw
more hardware at this instance since we are already committed to paying for that regardless.