Allow switch between Py2/3 virtualenvs
Prevents errors like

Python 2:
pkg_resources.DistributionNotFound: The 'importlib-metadata==3.1.0' distribution was not found and is required by commcare-cloud

Python 3
pkg_resources.DistributionNotFound: The 'ndg-httpsclient' distribution was not found and is required by commcare-cloud

Explanation: `pip install --editable .` writes package metadata, including a list of requirements, in the directory containing the "editable" dependency. This is a problem since requirements are different for Python 2 and 3. Errors happen when metadata for the opposite version of Python is present when running commcare-cloud commands. Removing requires.txt disables the run-time check of requirements by pkg_resources.