Move shard db pgbouncer to same host as pl/proxy
I found today that pl/proxy was routing its requests through pgbouncer on the same host
rather than the external host configured as the pgbouncer_host for the shard machines.
This means probably 99% of the traffic to the shard machines is already going through
pgbouncer3, so we should have the config reflect that. Right now if I update the pgbouncer
config on pgbouncer3, that will break traffic from pl/proxy to pg shard dbs