Disable storage autoscaling on pgmain0-production
This is to reflect the current live state. We found in our recent P1 that autoscaling storage
has a very dangerous side effect of, if it does not add enough during autoscaling,
preventing you from making any further changes for about 12 hours.
(That time probably depends on a lot of factors including how much data you have on disk.)