Multiple PgBouncer processes on a single machine
PgBouncer by default runs only one single-threaded process.
We implement an initial set of changes to allow us to run multiple
PgBouncer processes on a single machine, which share the same TCP
socket by using SO_REUSEPORT via the so_reuseport configuration
option.
The number of processes can be defined as a variable in the inventory
for the needed hosts, and defaults to 1.