Change pgbouncer tcp_keepidle to be less than AWS NLBs hard timeout
and increase tcp_keepcnt by the same factor.
The logic for lowering tcp_keepidle is that we need to send keepalive packets
to keep the NLB from timing out. The logic for raising tcp_keepcnt is to keep
the total length of time during which we are willing to send a keepalive the same.