refacor github api usage
This moves the code out of the `fab` package and into
the base commcare_cloud package.

It also makes use of generic prompts instead of supporting custom
prompts. Furthermore it will only prompt for credentials
if they are required either for private repos or for environments
that perform release tagging.