Note that they should look in to rate limiting
I don't think we actually have any documentation about this yet, so I'm
mostly putting this in here as a reminder.  Any project going through
this process will be working with Dimagi, so they can request Dimagi's
assistance with that, or we can build out docs on the feature when
needed.