Skip setup on --control for commands with no --branch arg
The assumption is that for commands that don't have a branch arg (django-manage, ssh, etc.)
the branch it's run on doesn't matter that much: that's why it doesn't use the --branch mechanism
to force the user to think about what branch they're on. For such commands, then,
pulling the latest code on the control machine and re-running the install steps
is an unneccessary waste of time.

Commands that do have the --branch option ('ansible-playbook', 'deploy', etc.)
have essentially declared that they are sensitive to the latest changes.
These commands will continue to trigger --control to do a full reset on the control environment
as it has been doing.