Always install the latest pgbouncer version
The pgbouncer package from postgresql.org is only available for the
latest version. We cannot set it to an arbitrary version as it would
break new installations.

The 'configure' tag allows us to perform normal operations without
upgrading the package, so let's set the version to 'latest' to avoid
manual interventions when we actually want to upgrade it.