copy-files - fix broken key removal
Previously it_rsa.tmp contained the key for one of the target hosts (probably the last one in the yml file); if there were multiple target hosts then only that last host key was cleaned up and all others remained in the authorized_keys file on the source host.

The number of run_ansible_module() calls was also reduced from 4 to 2 because the public key content is in the output of the "user" command with "generate_ssh_key=yes".