Remove scripts from target hosts on cleanup
Should make it more obvious if you happen to run a copy-files operation that stomps on someone else's concurrent operation (as opposed one that has long since been completed and cleaned up). Storing the scripts in a global location like /tmp/file_migration is not safe, but fixing that is out of scope for this cleanup.