Remove apply_patch command
We should not be modifying the running production code that way because there is no audit trail. Deploying a new release is a better way to do that. Also, it should not be needed anymore since all of our deployed servers have access to the internet.

The change to the "clean_releases" help text in commands.md is unexpected, but appears to have been caused by the removal of "apply_patch".