Fail deploy immediately when any task fails
Fabric also did this.

One example of why this is needed: staticfiles_pull_cache (task file executed on proxy[1:]) depends on shared files created by staticfiles_compress (task file executed on proxy[0]). The staticfiles_compress play will not match any hosts if an earlier task fails on proxy[0], but without this change Ansible would continue to execute tasks on other active proxy hosts. In general any task that updates a shared resource must run before later tasks that depend on that resource.

Another example that will be relevant when the _current symlink_ logic moves to Ansible: the current symlink should not be updated on a host if a task failed (effectively blocking that operation) on any other host.