Use 'diff: false' instead of 'no_log: true'
Avoid "censored": "the output has been hidden due to the fact that 'no_log: true' was specified for this result" in place of more informative output when task fails.

Note that some tasks (the "file" module, for example) do not respect 'diff: false', so 'no_log: true' has not been replace everywhere.