Rename 'git_mirrors' to '.git_mirrors'
Reduce the number of places that need to know about that directory:
- clean_releases and list-releases ignore hidden directories.
- people looking in the releases directory will be less likely to be
  confused because .git_mirrors is hidden from commands like `ls` by
  default.

This will have the side effect of breaking the git repositories in
existing releases when clean_releases is run the first time after this
is merged. The only instance where this would likely matter is in a
private release where someone was using git to update the release to a
new version of the code. Probably the easiest  workaround there is to
create a new private release and discard the old one. However, the
following commands can be used if a release gets broken that way and it
becomes necessary to (temporarily, until the next clean_releases run)
rehabilitate its git repo:

    sudo -u cchq ln -s .git_mirrors /home/cchq/www/<env>/releases/git_mirrors