Allow --check and be less strict about --diff
It's confusing to get a "that option is managed automatically" error when attempting to run with --check. The operator is forced to guess what will happen when lacking a thorough understanding of how the automatic option management works. It seems appropriate to perform a single --check run if that was requested by the operator.

The situation is similar for --diff, although a duplicate option does not cause harm and does not change the run mode.