Various documentation updates for data migration
There were a couple of defective links that I resolved as well as making
it more clear to readers that services should not be started back up
until after loading data into the new environment. I also replaced the
deploy step with a simple downtime end call because a deploy step is now
included when starting the steps for rebuilding an environment.