Decrease concurrency from 36 to 12
Having too much concurrency disrupts our ability to effectively rate
limit repeat records, since the more concurrency you have, the more
likely it is for a batch of slow tasks to be executed simultaneously
when usage is allowed. It isn't until enough of those slow tasks return
and report the request durations that rate limits apply.