Move submission-reprocessing-queue
Tasks on this queue can consume a lot of memory. Celerybeat is a smaller
machine than the other celery machines, and sometimes experiences OOMs
because of these tasks. We should not subject the celerybeat machine to
this since it is important for it to remain up for the sake of periodic
tasks.