Prevent unattended rabbitmq restart disrupting celery
lineinfile will probably work, although could fail under certain
conditions. For example, it will not add the blacklist item if an
unrelated line after the blacklist (a delimited range in the file)
matches the regex. Similarly it will add the item at the wrong place
(at the end of the file) if the `regexp` and `insertafter` patterns do
not match lines in the file, which seems possible if the file format
changed in a subtle way.