increase volume for es data nodes in b zones
we are going to copy the disks from production to c zone so we are not going to increase the disk size for that zone because when we will bring volumes we will unmount the ones that are attached. But for the b zone we will keep these disks as it is. And we will let ES replicate the data to these disks, for that these disks need to be of same size as in production