ignore exceptions thrown by django redis
In one of the issues that we faced, DNS for redis was updated which made redis unreachable to HQ. This caused HQ to display a 'Internal Server Error' instead of our custom Sad Danny. We have defined custom 500 handler but that is not triggered because django-redis error out before the handler is called. When we asked django redis to ignore the exceptions then if failed with a sad danny because the session handler was trying to get the session from redis and failed because cache was inaccessbile