Remove chunk sizes of 1
These were added on environments with a low volume of changes, which
caused pillows to wait a long time to create a chunk. A chunk size of 1
removed the need to wait, but recently the logic has improved to timeout
after waiting for 10 seconds, and process the chunk at that point.