Distinguish between legacy and drain modes
legacy - a celery machine continues to read from and write to the old
broker. This is to ensure remote control commands continue to function,
like shutdowns, but starts the process of draining the old broker since
all other commcare services read from and write to the new broker.

drain - finish draining by moving the celery machine to write to the new
broker as well. In this state, remote control commands will not work.